Override formatting for glossterm in admonitions
In admonitions, glossterms lack highlighting (background color and
dashed box around the term).

To fix that, use "!important" to override border and background-color,
which is necessary because div.notes uses "!important" to force
inheritance.

To keep the impact of this change limited to openstack, the change is
made in positioning-openstack.css. The downside of this solution is
that for openstack, custom.css is ignored with regards to these two
properties, and changing them there won't have any effect, which may
be confusing.

Related-Bug: 1303883

Change-Id: I989a41d61bdcddaac5734d5bd4b986465d8da443