Support db:idprefix, db:idfixup, and db:linkscope attributes on xi:include.
In some cases it is necessary to xi:include the same content more than
one time in the same document. This is difficult, however, when the
content contains xml:ids since this can result in duplicate xml:ids
and an invalid document. The db:idprefix, db:idfixup, and db:linkscope
attributes provide a means to pass instructions to the system
to allow it to munge the xml:id, linkend, and xlink:href attributes
in the xincluded content such that the id collisions do not occur.
For more information, see:

http://docbook.org/docs/transclusion/2010-12-09/transclusion.html#d6e179
https://github.com/docbook/transclusion-fixup
http://www.w3.org/TR/xinclude-11/

Change-Id: I90d1068a31a3e63f244f72f37b69b4ec8b13e79d