Allow error phrase to be listed with response code.
We usually list response codes, followed by an element name like so:

Error response codes: identityFault (400, 500, …), badRequest (400), unauthorized
(401), forbidden (403), badMethod (405), overLimit (413),
serviceUnavailable (503), itemNotFound (404)

The element name is the name of the element in the content.  For
example:

{
   "identityFault": {
        "message": "Fault",
        "details": "Error Details...",
        "code": 500
    }
}

This works great except that not all APIs follow this pattern.  This
patch allows for an extension in WADL that allows you to associate a
phrase wih a response. For example:

<response status="400" xmlns="http://wadl.dev.java.net/2009/02"
                       xmlns:rax="http://docs.rackspace.com/api"
                  rax:phrase="Bad Request">
...
</response>

which would display:

Bad Request (400)

Closes-Bug: #1338272

Change-Id: Ie5a998444b3fbfb6ad26b2b3ee32414443e0891d