Updated permissions for clouddocs-maven-plugin
After openstack-doc-core is added to the Reference permissions, it can
be removed from the clouddocs-maven-plugin-core group. The push
permissions allow me to work around a limitation in the code review
system, where it is impossible to merge two commits from a branch, and
force the creation of a merge commit (no fast-forward). This form is
used by the release process.

Change-Id: Ifdc04ebd8737c15e71f9d4fe4de997ce866b8820